Add files via upload
This is a task problem that requires each task to be performed by exactly one person. There are t tasks and p persons, the matrix representation of the problem shows rows as people and columns as tasks. The higher the number, the better the person performs the task. The solution to the problem is the assignment of persons to tasks (one task per person) that produces the best overall score.